[ALPS04011906] DISP: fix coverity
[Detail]
Coverity list:
  Bad bit shift operation:5459460
  Identical code for different branches:4323445
  Identical code for different branches:4323572
  Uninitialized scalar variable:8250106
  Unintentional integer overflow:8250179
  Unintentional integer overflow:8250433
  Dereference after null check:8260102
  Dereference after null check:8260110
  Dereference after null check:8260112
  Dereference after null check:8260127
  Dereference after null check:8260133
  Dereference after null check:8260136
  Dereference after null check:8260142
  Dereference after null check:8260143
  Dereference after null check:8260150
  Dereference after null check:8260154
  Dereference after null check:8260158
  Dereference after null check:8260169
  Dereference after null check:8260176
  Dereference after null check:8260186
  Dereference after null check:8260195
  Dereference after null check:8260197
  Dereference after null check:8260208
  Dereference after null check:8260209
  Dereference after null check:8260210
  Dereference after null check:8260221
  Dereference after null check:8260222
  Dereference after null check:8260232
  Dereference after null check:8260233
  Dereference after null check:8260237
  Dereference after null check:8260242
  Dereference after null check:8260245
  Dereference after null check:8260249
  Dereference after null check:8260259
  Dereference after null check:8260268
  Dereference after null check:8260275
  Dereference after null check:8260293
  Dereference after null check:8260297
  Dereference after null check:8260304
  Dereference after null check:8260312

MTK-Commit-Id: c22f5f3f126fa0d0f9728903cc9bd72a50043a2d

CR-Id: ALPS04011906
Feature: [Module]Display Driver

+++ drivers/misc/mediatek/video/mt6771/dispsys/ddp_rdma_ex.c
@@ -382,8 +382,8 @@ static void rdma_set_ultra_l(unsigned int idx, unsigned int bpp, void *handle,
 			fifo_off_ultra = 2;
 			fifo_off_dvfs = 2;
 		}
-		consume_rate = gsc->dst_width * gsc->dst_height	*
-			       if_fps * Bpp;
+		consume_rate = (unsigned long long)gsc->dst_width *
+				gsc->dst_height * if_fps * Bpp;
 		do_div(consume_rate, 1000);
 	} else {
 		fifo_off_drs_enter = 0;
@@ -391,8 +391,8 @@ static void rdma_set_ultra_l(unsigned int idx, unsigned int bpp, void *handle,
 		fifo_off_spm = 14; /* 10 times*/
 		fifo_off_dvfs = 2;
 		fifo_off_ultra = 0;
-		consume_rate = gsc->ext_dst_width * gsc->ext_dst_height *
-			       if_fps * Bpp;
+		consume_rate = (unsigned long long)gsc->ext_dst_width *
+				gsc->ext_dst_height * if_fps * Bpp;
 
 		do_div(consume_rate, 1000);
 	}
@@ -423,7 +423,7 @@ static void rdma_set_ultra_l(unsigned int idx, unsigned int bpp, void *handle,
 	issue_req_threshold = min(fifo_valid_size - preultra_high, 255U);
 
 	/* output valid should < total rdma data size, or hang will happen */
-	temp = gsc->rdma_width * gsc->rdma_height * Bpp;
+	temp = (unsigned long long)gsc->rdma_width * gsc->rdma_height * Bpp;
 	do_div(temp, 16);
 	temp -= 1;
 	output_valid_fifo_threshold = preultra_low < temp ? preultra_low : temp;
